starknet_api: change Proof from u32 to u8

+++ crates/apollo_protobuf/src/protobuf/protoc_output.rs
@@ -314,8 +314,8 @@ pub struct InvokeV3 {
 pub struct InvokeV3WithProof {
     #[prost(message, optional, tag = "1")]
     pub invoke: ::core::option::Option<InvokeV3>,
-    #[prost(uint32, repeated, tag = "2")]
-    pub proof: ::prost::alloc::vec::Vec<u32>,
+    #[prost(bytes = "vec", tag = "2")]
+    pub proof: ::prost::alloc::vec::Vec<u8>,
 }
 /// see <https://external.integration.starknet.io/feeder_gateway/get_transaction?transactionHash=0x29fd7881f14380842414cdfdd8d6c0b1f2174f8916edcfeb1ede1eb26ac3ef0>
 #[allow(clippy::derive_partial_eq_without_eq)]
